Refractor TreeViewComponent
