+++ src/components/TreeView/index.ts
@@ -1,0 +1,19 @@
+// Main component
+export { default as TreeViewComponent } from './TreeViewComponent';
+
+// Subcomponents
+export { default as TreeViewHeader } from './TreeViewHeader';
+export { default as TreeViewFilters } from './TreeViewFilters';
+export { default as TreeViewCanvas } from './TreeViewCanvas';
+export { default as NodeDetailsPanel } from './NodeDetailsPanel';
+export { default as TreeViewContextMenu } from './TreeViewContextMenu';
+export { default as TreeViewDeleteDialog } from './TreeViewDeleteDialog';
+
+// Hooks
+export { useTreeViewData } from './hooks/useTreeViewData';
+export { useTreeViewActions } from './hooks/useTreeViewActions';
+export { useTreeViewNodes } from './TreeViewNodes';
+export { useTreeViewEdges } from './TreeViewEdges';
+
+// Types
+export * from './types';
